Move serialization logic into graph.rs
This will simplify writing tests for table splitting, and is overall a
more reasonable separation of concerns.
